+++ server/src/main/java/org/opensearch/common/settings/ClusterSettings.java
@@ -839,6 +839,8 @@ public final class ClusterSettings extends AbstractScopedSettings {
                 BlobStoreRepository.SNAPSHOT_REPOSITORY_DATA_CACHE_THRESHOLD,
 
                 SearchService.CLUSTER_ALLOW_DERIVED_FIELD_SETTING,
+                SearchService.INNER_HITS_BATCH_ENABLED_SETTING,
+                SearchService.INNER_HITS_BATCH_SIZE_SETTING,
                 SearchService.QUERY_REWRITING_ENABLED_SETTING,
                 SearchService.QUERY_REWRITING_TERMS_THRESHOLD_SETTING,
                 FlushModeResolver.STREAMING_MAX_ESTIMATED_BUCKET_COUNT,